+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/DescriptorConverter.scala
@@ -1,0 +1,44 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.api.benchtool
+
+import com.daml.ledger.api.benchtool.Config.StreamConfig
+import com.daml.ledger.api.benchtool.submission.CommandSubmitter
+import com.daml.ledger.test.model.Foo.{Foo1, Foo2, Foo3}
+
+object DescriptorConverter {
+
+  def streamDescriptorToConfig(
+      descriptor: StreamDescriptor,
+      submissionSummary: CommandSubmitter.SubmissionSummary,
+  ): StreamConfig = {
+    import scalaz.syntax.tag._
+    def templateStringToId(template: String) = template match {
+      case "Foo1" => Foo1.id.unwrap
+      case "Foo2" => Foo2.id.unwrap
+      case "Foo3" => Foo3.id.unwrap
+      case invalid => throw new RuntimeException(s"Invalid template: $invalid")
+    }
+    def partyFromObservers(party: String): String =
+      submissionSummary.observers
+        .map(_.unwrap)
+        .find(_.contains(party))
+        .getOrElse(throw new RuntimeException(s"Observer not found: $party"))
+
+    val filters = descriptor.filters.map { filter =>
+      partyFromObservers(filter.party) -> Some(filter.templates.map(templateStringToId))
+    }.toMap
+
+    descriptor.streamType match {
+      case StreamDescriptor.StreamType.ActiveContracts =>
+        Config.StreamConfig.ActiveContractsStreamConfig(
+          name = descriptor.name,
+          filters = filters,
+        )
+      case invalid =>
+        throw new RuntimeException(s"Invalid stream type: $invalid")
+    }
+  }
+
+}
